pyproject.toml
[project]
name = "dx-without-excel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.3.3",
    "marimo>=0.18.4",
    "streamlit>=1.52.1",
    "pyyaml>=6.0.0",
    "ruff>=0.14.9",
    "playwright>=1.57.0",
    "pytest>=9.0.2",
    "pytest-cov>=6.0.0",
    "pytest-playwright>=0.4.4",
    "pre-commit>=4.0.0",
]

[tool.ruff.lint]
select = ["ALL"]
fixable = ["ALL"]
ignore = [
    "E501",
    "INP001",
    "TRY301",
    "RUF003",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--cov=src",
    "--cov-report=html",
    "--cov-report=term-missing",
    "--cov-report=xml",
    "--cov-fail-under=80"
]

[tool.coverage.run]
omit = [
    "src/presentation/app.py",  # Streamlit app is excluded from coverage
    "*/__pycache__/*",
    "*/tests/*"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
]
